Struct aws_sdk_elastictranscoder::client::fluent_builders::CreateJob

pub struct CreateJob<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> CreateJob<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(self) -> Result<CreateJobOutput, SdkError<CreateJobError>> where
R::Policy: SmithyRetryPolicy<CreateJobInputOperationOutputAlias, CreateJobOutput, CreateJobError, CreateJobInputOperationRetryAlias>,

pub fn pipeline_id(self, inp: impl Into<String>) -> Self

The Id of the pipeline that you want Elastic Transcoder to use for transcoding. The pipeline determines several settings, including the Amazon S3 bucket from which Elastic Transcoder gets the files to transcode and the bucket into which Elastic Transcoder puts the transcoded files.

pub fn set_pipeline_id(self, input: Option<String>) -> Self

pub fn input(self, inp: JobInput) -> Self

A section of the request body that provides information about the file that is being transcoded.

pub fn set_input(self, input: Option<JobInput>) -> Self

pub fn inputs(self, inp: impl Into<JobInput>) -> Self

Appends an item to Inputs.

To override the contents of this collection use set_inputs.

A section of the request body that provides information about the files that are being transcoded.

pub fn set_inputs(self, input: Option<Vec<JobInput>>) -> Self

pub fn output(self, inp: CreateJobOutput) -> Self

A section of the request body that provides information about the transcoded (target) file. We strongly recommend that you use the Outputs syntax instead of the Output syntax.

pub fn set_output(self, input: Option<CreateJobOutput>) -> Self

pub fn outputs(self, inp: impl Into<CreateJobOutput>) -> Self

Appends an item to Outputs.

To override the contents of this collection use set_outputs.

A section of the request body that provides information about the transcoded (target) files. We recommend that you use the Outputs syntax instead of the Output syntax.

pub fn set_outputs(self, input: Option<Vec<CreateJobOutput>>) -> Self

pub fn output_key_prefix(self, inp: impl Into<String>) -> Self

The value, if any, that you want Elastic Transcoder to prepend to the names of all files that this job creates, including output files, thumbnails, and playlists.

pub fn set_output_key_prefix(self, input: Option<String>) -> Self

pub fn playlists(self, inp: impl Into<CreateJobPlaylist>) -> Self

Appends an item to Playlists.

To override the contents of this collection use set_playlists.

If you specify a preset in PresetId for which the value of Container is fmp4 (Fragmented MP4) or ts (MPEG-TS), Playlists contains information about the master playlists that you want Elastic Transcoder to create.

The maximum number of master playlists in a job is 30.

pub fn set_playlists(self, input: Option<Vec<CreateJobPlaylist>>) -> Self

pub fn user_metadata(self, k: impl Into<String>, v: impl Into<String>) -> Self

Adds a key-value pair to UserMetadata.

To override the contents of this collection use set_user_metadata.

User-defined metadata that you want to associate with an Elastic Transcoder job. You specify metadata in key/value pairs, and you can add up to 10 key/value pairs per job. Elastic Transcoder does not guarantee that key/value pairs are returned in the same order in which you specify them.

pub fn set_user_metadata(self, input: Option<HashMap<String, String>>) -> Self