Module aws_sdk_elastictranscoder::error

Modules

access_denied_exception

See AccessDeniedException

incompatible_version_exception

See IncompatibleVersionException

internal_service_exception

See InternalServiceException

limit_exceeded_exception

See LimitExceededException

resource_in_use_exception

See ResourceInUseException

resource_not_found_exception

See ResourceNotFoundException

validation_exception

See ValidationException

Structs

AccessDeniedException

General authentication failure. The request was not signed correctly.

CancelJobError

CreateJobError

CreatePipelineError

CreatePresetError

DeletePipelineError

DeletePresetError

IncompatibleVersionException

InternalServiceException

Elastic Transcoder encountered an unexpected exception while trying to fulfill the request.

LimitExceededException

Too many operations for a given AWS account. For example, the number of pipelines

ListJobsByPipelineError

ListJobsByStatusError

ListPipelinesError

ListPresetsError

ReadJobError

ReadPipelineError

ReadPresetError

ResourceInUseException

The resource you are attempting to change is in use. For example, you are attempting

ResourceNotFoundException

The requested resource does not exist or is not available. For example, the pipeline

TestRoleError

UpdatePipelineError

UpdatePipelineNotificationsError

UpdatePipelineStatusError

ValidationException

One or more required parameter values were not provided in the request.

Enums

CancelJobErrorKind

CreateJobErrorKind

CreatePipelineErrorKind

CreatePresetErrorKind

DeletePipelineErrorKind

DeletePresetErrorKind

ListJobsByPipelineErrorKind

ListJobsByStatusErrorKind

ListPipelinesErrorKind

ListPresetsErrorKind

ReadJobErrorKind

ReadPipelineErrorKind

ReadPresetErrorKind

TestRoleErrorKind

UpdatePipelineErrorKind

UpdatePipelineNotificationsErrorKind

UpdatePipelineStatusErrorKind