Module aws_sdk_elastictranscoder::model

Modules

artwork

See Artwork

audio_codec_options

See AudioCodecOptions

audio_parameters

See AudioParameters

caption_format

See CaptionFormat

caption_source

See CaptionSource

captions

See Captions

clip

See Clip

create_job_output

See CreateJobOutput

create_job_playlist

See CreateJobPlaylist

detected_properties

See DetectedProperties

encryption

See Encryption

hls_content_protection

See HlsContentProtection

input_captions

See InputCaptions

job

See Job

job_album_art

See JobAlbumArt

job_input

See JobInput

job_output

See JobOutput

job_watermark

See JobWatermark

notifications

See Notifications

permission

See Permission

pipeline

See Pipeline

pipeline_output_config

See PipelineOutputConfig

play_ready_drm

See PlayReadyDrm

playlist

See Playlist

preset

See Preset

preset_watermark

See PresetWatermark

thumbnails

See Thumbnails

time_span

See TimeSpan

timing

See Timing

video_parameters

See VideoParameters

warning

See Warning

Structs

Artwork

The file to be used as album art. There can be multiple artworks associated with an

AudioCodecOptions

Options associated with your audio codec.

AudioParameters

Parameters required for transcoding audio.

CaptionFormat

The file format of the output captions. If you leave this value blank, Elastic Transcoder returns an error.

CaptionSource

A source file for the input sidecar captions used during the transcoding process.

Captions

The captions to be created, if any.

Clip

Settings for one clip in a composition. All jobs in a playlist must have the same clip settings.

CreateJobOutput

The CreateJobOutput structure.

CreateJobPlaylist

Information about the master playlist.

DetectedProperties

The detected properties of the input file. Elastic Transcoder identifies these values from the input file.

Encryption

The encryption settings, if any, that are used for decrypting your input files or

HlsContentProtection

The HLS content protection settings, if any, that you want Elastic Transcoder to apply to your output files.

InputCaptions

The captions to be created, if any.

Job

A section of the response body that provides information about the job that is created.

JobAlbumArt

The .jpg or .png file associated with an audio file.

JobInput

Information about the file that you're transcoding.

JobOutput

JobWatermark

Watermarks can be in .png or .jpg format. If you want to display a watermark that is not

Notifications

The Amazon Simple Notification Service (Amazon SNS) topic or topics to notify in order to report job status.

Permission

The Permission structure.

Pipeline

The pipeline (queue) that is used to manage jobs.

PipelineOutputConfig

The PipelineOutputConfig structure.

PlayReadyDrm

The PlayReady DRM settings, if any, that you want Elastic Transcoder to apply to the output files associated with this playlist.

Playlist

Use Only for Fragmented MP4 or MPEG-TS Outputs. If you specify a preset for which the value of Container

Preset

Presets are templates that contain most of the settings for transcoding media files

PresetWatermark

Settings for the size, location, and opacity of graphics that you want Elastic Transcoder to

Thumbnails

Thumbnails for videos.

TimeSpan

Settings that determine when a clip begins and how long it lasts.

Timing

Details about the timing of a job.

VideoParameters

The VideoParameters structure.

Warning

Elastic Transcoder returns a warning if the resources used by your pipeline are not in the same region