Struct aws_sdk_elastictranscoder::model::job_output::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for JobOutput

Implementations

impl Builder

pub fn id(self, input: impl Into<String>) -> Self

A sequential counter, starting with 1, that identifies an output among the outputs from the current job. In the Output syntax, this value is always 1.

pub fn set_id(self, input: Option<String>) -> Self

pub fn key(self, input: impl Into<String>) -> Self

The name to assign to the transcoded file. Elastic Transcoder saves the file in the Amazon S3 bucket specified by the OutputBucket object in the pipeline that is specified by the pipeline ID.

pub fn set_key(self, input: Option<String>) -> Self

pub fn thumbnail_pattern(self, input: impl Into<String>) -> Self

Whether you want Elastic Transcoder to create thumbnails for your videos and, if so, how you want Elastic Transcoder to name the files.

If you don't want Elastic Transcoder to create thumbnails, specify "".

If you do want Elastic Transcoder to create thumbnails, specify the information that you want to include in the file name for each thumbnail. You can specify the following values in any sequence:

{count} (Required): If you want to create thumbnails, you must include {count} in the ThumbnailPattern object. Wherever you specify {count}, Elastic Transcoder adds a five-digit sequence number (beginning with 00001) to thumbnail file names. The number indicates where a given thumbnail appears in the sequence of thumbnails for a transcoded file.

If you specify a literal value and/or {resolution} but you omit {count}, Elastic Transcoder returns a validation error and does not create the job.

Literal values (Optional): You can specify literal values anywhere in the ThumbnailPattern object. For example, you can include them as a file name prefix or as a delimiter between {resolution} and {count}.

{resolution} (Optional): If you want Elastic Transcoder to include the resolution in the file name, include {resolution} in the ThumbnailPattern object.

When creating thumbnails, Elastic Transcoder automatically saves the files in the format (.jpg or .png) that appears in the preset that you specified in the PresetID value of CreateJobOutput. Elastic Transcoder also appends the applicable file name extension.

pub fn set_thumbnail_pattern(self, input: Option<String>) -> Self

pub fn thumbnail_encryption(self, input: Encryption) -> Self

The encryption settings, if any, that you want Elastic Transcoder to apply to your thumbnail.

pub fn set_thumbnail_encryption(self, input: Option<Encryption>) -> Self

pub fn rotate(self, input: impl Into<String>) -> Self

The number of degrees clockwise by which you want Elastic Transcoder to rotate the output relative to the input. Enter one of the following values:

auto, 0, 90, 180, 270

The value auto generally works only if the file that you're transcoding contains rotation metadata.

pub fn set_rotate(self, input: Option<String>) -> Self

pub fn preset_id(self, input: impl Into<String>) -> Self

The value of the Id object for the preset that you want to use for this job. The preset determines the audio, video, and thumbnail settings that Elastic Transcoder uses for transcoding. To use a preset that you created, specify the preset ID that Elastic Transcoder returned in the response when you created the preset. You can also use the Elastic Transcoder system presets, which you can get with ListPresets.

pub fn set_preset_id(self, input: Option<String>) -> Self

pub fn segment_duration(self, input: impl Into<String>) -> Self

(Outputs in Fragmented MP4 or MPEG-TS format only.

If you specify a preset in PresetId for which the value of Container is fmp4 (Fragmented MP4) or ts (MPEG-TS), SegmentDuration is the target maximum duration of each segment in seconds. For HLSv3 format playlists, each media segment is stored in a separate .ts file. For HLSv4, MPEG-DASH, and Smooth playlists, all media segments for an output are stored in a single file. Each segment is approximately the length of the SegmentDuration, though individual segments might be shorter or longer.

The range of valid values is 1 to 60 seconds. If the duration of the video is not evenly divisible by SegmentDuration, the duration of the last segment is the remainder of total length/SegmentDuration.

Elastic Transcoder creates an output-specific playlist for each output HLS output that you specify in OutputKeys. To add an output to the master playlist for this job, include it in the OutputKeys of the associated playlist.

pub fn set_segment_duration(self, input: Option<String>) -> Self

pub fn status(self, input: impl Into<String>) -> Self

The status of one output in a job. If you specified only one output for the job, Outputs:Status is always the same as Job:Status. If you specified more than one output:

Job:Status and Outputs:Status for all of the outputs is Submitted until Elastic Transcoder starts to process the first output.

When Elastic Transcoder starts to process the first output, Outputs:Status for that output and Job:Status both change to Progressing. For each output, the value of Outputs:Status remains Submitted until Elastic Transcoder starts to process the output.

Job:Status remains Progressing until all of the outputs reach a terminal status, either Complete or Error.

When all of the outputs reach a terminal status, Job:Status changes to Complete only if Outputs:Status for all of the outputs is Complete. If Outputs:Status for one or more outputs is Error, the terminal status for Job:Status is also Error.

The value of Status is one of the following: Submitted, Progressing, Complete, Canceled, or Error.

pub fn set_status(self, input: Option<String>) -> Self

pub fn status_detail(self, input: impl Into<String>) -> Self

Information that further explains Status.

pub fn set_status_detail(self, input: Option<String>) -> Self

pub fn duration(self, input: i64) -> Self

Duration of the output file, in seconds.

pub fn set_duration(self, input: Option<i64>) -> Self

pub fn width(self, input: i32) -> Self

Specifies the width of the output file in pixels.

pub fn set_width(self, input: Option<i32>) -> Self

pub fn height(self, input: i32) -> Self

Height of the output file, in pixels.

pub fn set_height(self, input: Option<i32>) -> Self

pub fn frame_rate(self, input: impl Into<String>) -> Self

Frame rate of the output file, in frames per second.

pub fn set_frame_rate(self, input: Option<String>) -> Self

pub fn file_size(self, input: i64) -> Self

File size of the output file, in bytes.

pub fn set_file_size(self, input: Option<i64>) -> Self

pub fn duration_millis(self, input: i64) -> Self

Duration of the output file, in milliseconds.

pub fn set_duration_millis(self, input: Option<i64>) -> Self

pub fn watermarks(self, input: impl Into<JobWatermark>) -> Self

pub fn set_watermarks(self, input: Option<Vec<JobWatermark>>) -> Self

pub fn album_art(self, input: JobAlbumArt) -> Self

The album art to be associated with the output file, if any.

pub fn set_album_art(self, input: Option<JobAlbumArt>) -> Self

pub fn composition(self, input: impl Into<Clip>) -> Self

pub fn set_composition(self, input: Option<Vec<Clip>>) -> Self

pub fn captions(self, input: Captions) -> Self

You can configure Elastic Transcoder to transcode captions, or subtitles, from one format to another. All captions must be in UTF-8. Elastic Transcoder supports two types of captions:

Embedded: Embedded captions are included in the same file as the audio and video. Elastic Transcoder supports only one embedded caption per language, to a maximum of 300 embedded captions per file.

Valid input values include: CEA-608 (EIA-608, first non-empty channel only), CEA-708 (EIA-708, first non-empty channel only), and mov-text

Valid outputs include: mov-text

Elastic Transcoder supports a maximum of one embedded format per output.

Sidecar: Sidecar captions are kept in a separate metadata file from the audio and video data. Sidecar captions require a player that is capable of understanding the relationship between the video file and the sidecar file. Elastic Transcoder supports only one sidecar caption per language, to a maximum of 20 sidecar captions per file.

Valid input values include: dfxp (first div element only), ebu-tt, scc, smpt, srt, ttml (first div element only), and webvtt

Valid outputs include: dfxp (first div element only), scc, srt, and webvtt.

If you want ttml or smpte-tt compatible captions, specify dfxp as your output format.

Elastic Transcoder does not support OCR (Optical Character Recognition), does not accept pictures as a valid input for captions, and is not available for audio-only transcoding. Elastic Transcoder does not preserve text formatting (for example, italics) during the transcoding process.

To remove captions or leave the captions empty, set Captions to null. To pass through existing captions unchanged, set the MergePolicy to MergeRetain, and pass in a null CaptionSources array.

For more information on embedded files, see the Subtitles Wikipedia page.

For more information on sidecar files, see the Extensible Metadata Platform and Sidecar file Wikipedia pages.

pub fn set_captions(self, input: Option<Captions>) -> Self

pub fn encryption(self, input: Encryption) -> Self

The encryption settings, if any, that you want Elastic Transcoder to apply to your output files. If you choose to use encryption, you must specify a mode to use. If you choose not to use encryption, Elastic Transcoder writes an unencrypted file to your Amazon S3 bucket.

pub fn set_encryption(self, input: Option<Encryption>) -> Self

pub fn applied_color_space_conversion(self, input: impl Into<String>) -> Self

If Elastic Transcoder used a preset with a ColorSpaceConversionMode to transcode the output file, the AppliedColorSpaceConversion parameter shows the conversion used. If no ColorSpaceConversionMode was defined in the preset, this parameter is not be included in the job response.

pub fn set_applied_color_space_conversion(self, input: Option<String>) -> Self

pub fn build(self) -> JobOutput

Consumes the builder and constructs a JobOutput