Struct aws_sdk_elastictranscoder::model::video_parameters::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for VideoParameters

Implementations

impl Builder

pub fn codec(self, input: impl Into<String>) -> Self

The video codec for the output file. Valid values include gif, H.264, mpeg2, vp8, and vp9. You can only specify vp8 and vp9 when the container type is webm, gif when the container type is gif, and mpeg2 when the container type is mpg.

pub fn set_codec(self, input: Option<String>) -> Self

pub fn codec_options(self, k: impl Into<String>, v: impl Into<String>) -> Self

pub fn set_codec_options(self, input: Option<HashMap<String, String>>) -> Self

pub fn keyframes_max_dist(self, input: impl Into<String>) -> Self

Applicable only when the value of Video:Codec is one of H.264, MPEG2, or VP8.

The maximum number of frames between key frames. Key frames are fully encoded frames; the frames between key frames are encoded based, in part, on the content of the key frames. The value is an integer formatted as a string; valid values are between 1 (every frame is a key frame) and 100000, inclusive. A higher value results in higher compression but may also discernibly decrease video quality.

For Smooth outputs, the FrameRate must have a constant ratio to the KeyframesMaxDist. This allows Smooth playlists to switch between different quality levels while the file is being played.

For example, an input file can have a FrameRate of 30 with a KeyframesMaxDist of 90. The output file then needs to have a ratio of 1:3. Valid outputs would have FrameRate of 30, 25, and 10, and KeyframesMaxDist of 90, 75, and 30, respectively.

Alternately, this can be achieved by setting FrameRate to auto and having the same values for MaxFrameRate and KeyframesMaxDist.

pub fn set_keyframes_max_dist(self, input: Option<String>) -> Self

pub fn fixed_gop(self, input: impl Into<String>) -> Self

Applicable only when the value of Video:Codec is one of H.264, MPEG2, or VP8.

Whether to use a fixed value for FixedGOP. Valid values are true and false:

true: Elastic Transcoder uses the value of KeyframesMaxDist for the distance between key frames (the number of frames in a group of pictures, or GOP).

false: The distance between key frames can vary.

FixedGOP must be set to true for fmp4 containers.

pub fn set_fixed_gop(self, input: Option<String>) -> Self

pub fn bit_rate(self, input: impl Into<String>) -> Self

The bit rate of the video stream in the output file, in kilobits/second. Valid values depend on the values of Level and Profile. If you specify auto, Elastic Transcoder uses the detected bit rate of the input source. If you specify a value other than auto, we recommend that you specify a value less than or equal to the maximum H.264-compliant value listed for your level and profile:

Level - Maximum video bit rate in kilobits/second (baseline and main Profile) : maximum video bit rate in kilobits/second (high Profile)

1 - 64 : 80

1b - 128 : 160

1.1 - 192 : 240

1.2 - 384 : 480

1.3 - 768 : 960

2 - 2000 : 2500

3 - 10000 : 12500

3.1 - 14000 : 17500

3.2 - 20000 : 25000

4 - 20000 : 25000

4.1 - 50000 : 62500

pub fn set_bit_rate(self, input: Option<String>) -> Self

pub fn frame_rate(self, input: impl Into<String>) -> Self

The frames per second for the video stream in the output file. Valid values include:

auto, 10, 15, 23.97, 24, 25, 29.97, 30, 60

If you specify auto, Elastic Transcoder uses the detected frame rate of the input source. If you specify a frame rate, we recommend that you perform the following calculation:

Frame rate = maximum recommended decoding speed in luma samples/second / (width in pixels * height in pixels)

where:

width in pixels and height in pixels represent the Resolution of the output video.

maximum recommended decoding speed in Luma samples/second is less than or equal to the maximum value listed in the following table, based on the value that you specified for Level.

The maximum recommended decoding speed in Luma samples/second for each level is described in the following list (Level - Decoding speed):

1 - 380160

1b - 380160

1.1 - 76800

1.2 - 1536000

1.3 - 3041280

2 - 3041280

2.1 - 5068800

2.2 - 5184000

3 - 10368000

3.1 - 27648000

3.2 - 55296000

4 - 62914560

4.1 - 62914560

pub fn set_frame_rate(self, input: Option<String>) -> Self

pub fn max_frame_rate(self, input: impl Into<String>) -> Self

If you specify auto for FrameRate, Elastic Transcoder uses the frame rate of the input video for the frame rate of the output video. Specify the maximum frame rate that you want Elastic Transcoder to use when the frame rate of the input video is greater than the desired maximum frame rate of the output video. Valid values include: 10, 15, 23.97, 24, 25, 29.97, 30, 60.

pub fn set_max_frame_rate(self, input: Option<String>) -> Self

pub fn resolution(self, input: impl Into<String>) -> Self

To better control resolution and aspect ratio of output videos, we recommend that you use the values MaxWidth, MaxHeight, SizingPolicy, PaddingPolicy, and DisplayAspectRatio instead of Resolution and AspectRatio. The two groups of settings are mutually exclusive. Do not use them together.

The width and height of the video in the output file, in pixels. Valid values are auto and width x height:

auto: Elastic Transcoder attempts to preserve the width and height of the input file, subject to the following rules.

width x height : The width and height of the output video in pixels.

Note the following about specifying the width and height:

The width must be an even integer between 128 and 4096, inclusive.

The height must be an even integer between 96 and 3072, inclusive.

If you specify a resolution that is less than the resolution of the input file, Elastic Transcoder rescales the output file to the lower resolution.

If you specify a resolution that is greater than the resolution of the input file, Elastic Transcoder rescales the output to the higher resolution.

We recommend that you specify a resolution for which the product of width and height is less than or equal to the applicable value in the following list (List - Max width x height value):

1 - 25344

1b - 25344

1.1 - 101376

1.2 - 101376

1.3 - 101376

2 - 101376

2.1 - 202752

2.2 - 404720

3 - 404720

3.1 - 921600

3.2 - 1310720

4 - 2097152

4.1 - 2097152

pub fn set_resolution(self, input: Option<String>) -> Self

pub fn aspect_ratio(self, input: impl Into<String>) -> Self

To better control resolution and aspect ratio of output videos, we recommend that you use the values MaxWidth, MaxHeight, SizingPolicy, PaddingPolicy, and DisplayAspectRatio instead of Resolution and AspectRatio. The two groups of settings are mutually exclusive. Do not use them together.

The display aspect ratio of the video in the output file. Valid values include:

auto, 1:1, 4:3, 3:2, 16:9

If you specify auto, Elastic Transcoder tries to preserve the aspect ratio of the input file.

If you specify an aspect ratio for the output file that differs from aspect ratio of the input file, Elastic Transcoder adds pillarboxing (black bars on the sides) or letterboxing (black bars on the top and bottom) to maintain the aspect ratio of the active region of the video.

pub fn set_aspect_ratio(self, input: Option<String>) -> Self

pub fn max_width(self, input: impl Into<String>) -> Self

The maximum width of the output video in pixels. If you specify auto, Elastic Transcoder uses 1920 (Full HD) as the default value. If you specify a numeric value, enter an even integer between 128 and 4096.

pub fn set_max_width(self, input: Option<String>) -> Self

pub fn max_height(self, input: impl Into<String>) -> Self

The maximum height of the output video in pixels. If you specify auto, Elastic Transcoder uses 1080 (Full HD) as the default value. If you specify a numeric value, enter an even integer between 96 and 3072.

pub fn set_max_height(self, input: Option<String>) -> Self

pub fn display_aspect_ratio(self, input: impl Into<String>) -> Self

The value that Elastic Transcoder adds to the metadata in the output file.

pub fn set_display_aspect_ratio(self, input: Option<String>) -> Self

pub fn sizing_policy(self, input: impl Into<String>) -> Self

Specify one of the following values to control scaling of the output video:

Fit: Elastic Transcoder scales the output video so it matches the value that you specified in either MaxWidth or MaxHeight without exceeding the other value.

Fill: Elastic Transcoder scales the output video so it matches the value that you specified in either MaxWidth or MaxHeight and matches or exceeds the other value. Elastic Transcoder centers the output video and then crops it in the dimension (if any) that exceeds the maximum value.

Stretch: Elastic Transcoder stretches the output video to match the values that you specified for MaxWidth and MaxHeight. If the relative proportions of the input video and the output video are different, the output video will be distorted.

Keep: Elastic Transcoder does not scale the output video. If either dimension of the input video exceeds the values that you specified for MaxWidth and MaxHeight, Elastic Transcoder crops the output video.

ShrinkToFit: Elastic Transcoder scales the output video down so that its dimensions match the values that you specified for at least one of MaxWidth and MaxHeight without exceeding either value. If you specify this option, Elastic Transcoder does not scale the video up.

ShrinkToFill: Elastic Transcoder scales the output video down so that its dimensions match the values that you specified for at least one of MaxWidth and MaxHeight without dropping below either value. If you specify this option, Elastic Transcoder does not scale the video up.

pub fn set_sizing_policy(self, input: Option<String>) -> Self

pub fn padding_policy(self, input: impl Into<String>) -> Self

When you set PaddingPolicy to Pad, Elastic Transcoder may add black bars to the top and bottom and/or left and right sides of the output video to make the total size of the output video match the values that you specified for MaxWidth and MaxHeight.

pub fn set_padding_policy(self, input: Option<String>) -> Self

pub fn watermarks(self, input: impl Into<PresetWatermark>) -> Self

pub fn set_watermarks(self, input: Option<Vec<PresetWatermark>>) -> Self

pub fn build(self) -> VideoParameters

Consumes the builder and constructs a VideoParameters