Module aws_sdk_elastictranscoder::operation

Structs

CancelJob

The CancelJob operation cancels an unfinished job.

CreateJob

When you create a job, Elastic Transcoder returns JSON data that includes the values that you specified

CreatePipeline

The CreatePipeline operation creates a pipeline with settings that you specify.

CreatePreset

The CreatePreset operation creates a preset with settings that you specify.

DeletePipeline

The DeletePipeline operation removes a pipeline.

DeletePreset

The DeletePreset operation removes a preset that you've added in an AWS region.

ListJobsByPipeline

The ListJobsByPipeline operation gets a list of the jobs currently in a pipeline.

ListJobsByStatus

The ListJobsByStatus operation gets a list of jobs that have a specified status. The response

ListPipelines

The ListPipelines operation gets a list of the pipelines associated with the current AWS account.

ListPresets

The ListPresets operation gets a list of the default presets included with Elastic Transcoder and the presets that

ReadJob

The ReadJob operation returns detailed information about a job.

ReadPipeline

The ReadPipeline operation gets detailed information about a pipeline.

ReadPreset

The ReadPreset operation gets detailed information about a preset.

TestRole

The TestRole operation tests the IAM role used to create the pipeline.

UpdatePipeline

Use the UpdatePipeline operation to update settings for a pipeline.

UpdatePipelineNotifications

With the UpdatePipelineNotifications operation, you can update Amazon Simple Notification Service (Amazon SNS) notifications for a pipeline.

UpdatePipelineStatus

The UpdatePipelineStatus operation pauses or reactivates a pipeline, so that the pipeline