Module aws_sdk_elastictranscoder::output

Modules

cancel_job_output

See CancelJobOutput

create_job_output

See CreateJobOutput

create_pipeline_output

See CreatePipelineOutput

create_preset_output

See CreatePresetOutput

delete_pipeline_output

See DeletePipelineOutput

delete_preset_output

See DeletePresetOutput

list_jobs_by_pipeline_output

See ListJobsByPipelineOutput

list_jobs_by_status_output

See ListJobsByStatusOutput

list_pipelines_output

See ListPipelinesOutput

list_presets_output

See ListPresetsOutput

read_job_output

See ReadJobOutput

read_pipeline_output

See ReadPipelineOutput

read_preset_output

See ReadPresetOutput

test_role_output

See TestRoleOutput

update_pipeline_notifications_output

See UpdatePipelineNotificationsOutput

update_pipeline_output

See UpdatePipelineOutput

update_pipeline_status_output

See UpdatePipelineStatusOutput

Structs

CancelJobOutput

The response body contains a JSON object. If the job is successfully canceled, the value

CreateJobOutput

The CreateJobResponse structure.

CreatePipelineOutput

When you create a pipeline, Elastic Transcoder returns the values that you specified in the request.

CreatePresetOutput

The CreatePresetResponse structure.

DeletePipelineOutput

The DeletePipelineResponse structure.

DeletePresetOutput

The DeletePresetResponse structure.

ListJobsByPipelineOutput

The ListJobsByPipelineResponse structure.

ListJobsByStatusOutput

ListPipelinesOutput

A list of the pipelines associated with the current AWS account.

ListPresetsOutput

The ListPresetsResponse structure.

ReadJobOutput

The ReadJobResponse structure.

ReadPipelineOutput

The ReadPipelineResponse structure.

ReadPresetOutput

The ReadPresetResponse structure.

TestRoleOutput

The TestRoleResponse structure.

UpdatePipelineNotificationsOutput

The UpdatePipelineNotificationsResponse structure.

UpdatePipelineOutput

When you update a pipeline, Elastic Transcoder returns the values that you specified in the request.

UpdatePipelineStatusOutput

When you update status for a pipeline,